backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - MongoDB
motor==3.3.2  # Async MongoDB driver for FastAPI
pymongo==4.6.0
beanie==1.24.0  # ODM for MongoDB with Pydantic

# Caching & Background Tasks
redis>=4.5.2,<5.0.0
celery[redis]==5.3.4

# HTTP & Scraping
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2  # for dynamic content

# ML & NLP
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
numpy==1.25.2
pandas==2.1.3
textblob==0.17.1
nltk==3.8.1

# Data Processing
joblib==1.3.2
python-multipart==0.0.6

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # for testing FastAPI

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0